Rewrite serialization script (#49)
Cleaner and correct handling of what the expected outcomes can be.
This is just a test script, so we don't need a careful domain-specific
review.

I only expect `.nt`, `.nq` or `.ttl` files as the inputs.
I only expect one `.jelly` file as the output.

Single NT/TTL files are serialized as flat triples, NQ files are
absorbed to a dataset and serialized as named graphs or quads (depending
on the `--graphs` option). Multiple NT/TTL are merged as well and the
graphs are named as their filenames, as well serialized as named graphs
or quads. NT/TTL and NQ files can be mixed.