Make CI pass – remove towncrier (part 2) (#147)
Issue #139 

I looked into this "towncrier" thing and it seems to be a tool intended
for very complex, very busy projects, where keeping a changelog is very
hard using traditional means. As far as I understand, to cope with that,
it adds extra responsibilities to PR authors and project maintainers...
but these are not documented:
https://jelly-rdf.github.io/pyjelly/dev/contributing/#contributions so I
have no idea what I should do to make the damn thing happy.

Simple is better, so I say let's just remove it and have a changelog on
the releases page, like in Jelly-JVM and jelly-cli.